Code Refinement
Here I removed as much code and functions as I could, hovever I did not
manage to free up as much memory as I would have liked. This is now very
light weight.
I also have it setup to send a test packet on request. This packet is in
the correct format for our design.