Stabilised Network Connection
The device used to be unstable, in that it would go into an infinite
loop if it had to wait > 1min for a recieved message, or if a connecion
was broken while waiting for a message. I have fixed this by checking
for the bytes recieved, which gets allocated -57 (socket error value) if
a connection breaks.
The current send rate is 166,04 kS/s. Ha ha ha. Awesome! Almost no need
to implement the dubble buffer system.