Toggle/Dual DMA operational!
This is great. I finnaly managed to fix this. The problem with my last
attempt (2 days work) is that the DMA was sending a full word! IE 2
bytes. Therefore the second buffer would overflow into the interrupt
memory area, and shut the system down! Anyway, this is now working. Not
sue about the speed improvement... will have to wait for Manie. Code is
a bit untidy, but ok. Rearranged how the main loop operates.