Missed ADC cycle measurement
This code is stable. I implemented an interrupt to count the ammount of
ADC cycles lost while waiting for the send to finish. Also wrote the
code to insert that value into the packet header. the speed with this
check is only +-160 kS/s. Without the check the speed is 197kS/s. This
is a big drop. May be worth excluding this information, Otherwise I must
speed the system up. less instructions, fewer samples, slower sample
rates.
Note: There is a connection issue. If a stream of data is
interrupted/broken, then the next connection will break. No clue why it
does that.