ADC Fixed
This is almost the completed code. The ADC was not working. It turns out
that it does not like using MCLK as the source. I changed it back to
MODCLK (slower I think) and it worked fine. I set clock cycles to 4, and
it semi worked. Strangely enough the process actually got faster!!!
There was a bug in that the ADC would staop working after the first use.
I think this may be ixed, because I set the ADC clock cycles to 16.
Still needs lots of debugging, but we are doing well. Managed to send a
live sine wave at 6 kHz!!!