Selectable channels
The code is looking a bit beter. Still not sure about the stability of
the MSP. Seems to act randomly. Need to test more. I have included
functions to clear the ADC and DMA at regular intervals. Hopefully this
will help. I have also made the devic configerable, so that different
channels can be selected. Need to configure for three channels. Will be
interesting for Manie to deal with.